requirements.txt
Babel==2.6.0
Django==2.0.7
django-ckeditor==5.6.1
django-debug-toolbar==1.9.1
django-js-asset==1.1.0
django-mptt==0.9.0
django-phone-field==1.5.4
django-phonenumber-field==2.0.0
phonenumberslite==8.9.10
Pillow==5.2.0
pytz==2018.5
sqlparse==0.2.4
